eval_5/requirements.txt
absl-py==1.0.0
ai2thor==2.5.0
appnope==0.1.3
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
aws-requests-auth==0.4.3
backcall==0.2.0
botocore==1.24.37
cachetools==5.1.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colour==0.1.5
cycler==0.10.0
debugpy==1.6.0
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.1
flatbuffers==1.12
fonttools==4.32.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==3.6.0
idna==2.10
imageio==2.16.1
import-ipynb==0.1.4
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.12.1
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jmespath==1.0.0
jsonschema==4.6.0
jupyter-client==7.2.2
jupyter-core==4.9.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==14.0.1
machine-common-sense==0.5.7
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
msgpack==1.0.3
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
numpy==1.22.3
numpyencoder==0.3.0
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
progressbar2==4.0.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
Pygments==2.11.2
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-utils==3.1.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
roboflow==0.2.4
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.5.12
setproctitle==1.2.3
Shapely==1.8.1.post1
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2022.4.8
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typeguard==2.13.3
typing_extensions==4.1.1
urllib3==1.26.6
wandb==0.12.16
wcwidth==0.2.5
Werkzeug==2.1.1
wget==3.2
wrapt==1.14.1
zipp==3.8.0
